fix: add base exception
